Split the test for the entry/grouped entry handler into two files
